[US_TN] Compliant Reporting TES Refactor - Compile final criteria  (Recidiviz/recidiviz-data#21341)
## Description of the change

Should be merged in after Recidiviz/recidiviz-data#21057. Builds on @agaidus's PR to add the
relevant offense-type related criteria to the `no_discretion` and
`with_discretion` tasks. Also does some clean up of the
`with_discretion` task to re-use all the "base" required criteria and
add a single criteria, `ineligible_without_discretion`. This makes the
two tasks explicitly mutually exclusive and allows the `with_discretion`
task to be a catch all for anyone that is considered ineligible because
of offense types, missing sentence information, and zero tolerance codes
(s/o to @samanthahuff for the suggestion and help brainstorming!!)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18175

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3d8b3052880afdc7b78a8338877772736a87c67b