[Airflow] Fix typing annotation. Airflow environment does not use Python 3.9 (Recidiviz/recidiviz-data#21563)
## Description of the change

This fixes the DAG import errors in staging.
```python
Broken DAG: [/home/airflow/gcs/dags/calculation_dag.py] Traceback (most recent call last):
  File "/home/airflow/gcs/dags/calculation_dag.py", line 56, in <module>
    from recidiviz.airflow.dags.utils.pagerduty_integration import (
  File "/home/airflow/gcs/dags/recidiviz/airflow/dags/utils/pagerduty_integration.py", line 26, in <module>
    def pagerduty_integration_email(conn_id: str) -> Optional[list[str]]:
TypeError: 'type' object is not subscriptable
```

It also updates our CI checks to use Python 3.8.12 which is what we run
in Composer.

https://cloud.google.com/composer/docs/concepts/versioning/composer-versions#images

A failing test suite that shows the same error as was seen in staging
can be found here-

https://github.com/Recidiviz/recidiviz-data/actions/runs/5272332541/jobs/9534458091

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: 6e84bda4ca8e9f2a4d596995a37c5ed220f12632