Remove gitignored .code-workspace file (Recidiviz/recidiviz-data#21469)
## Description of the change

We [gitignore .code-workspace
files](https://github.com/Recidiviz/recidiviz-data/blob/f0639ec2b2a7c12316bd75fffad51a2329475909/.gitignore#L188)
but have one in our repo, which this PR removes. I found this because
I'd like my .code-workspace file (set by VSCode) to be a bit different
than what's in the repo, but that's based on my personal file structure
and should not affect the rest of the team, so I don't think that file
should be committed to the repo.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 42b097fee4a1445b25a04135bf64adf622635857