[US_MO] Overdue for Hearing (Recidiviz/recidiviz-data#21423)
## Description of the change

Updates the Restrictive Housing Hearings criteria and opportunity record
query to surface people who are overdue for hearings, and those who are
in Restrictive Housing but do not have an upcoming or overdue hearing.

The `HAS_UPCOMING_HEARING` criterion is replaced with
`OVERDUE_FOR_HEARING`, changing the definition of eligibility
accordingly. Those who were previously eligible for
`HAS_UPCOMING_HEARING` are now ineligible, but will be surfaced as
Upcoming by being `IN_RESTRICTIVE_HOUSING` and within 7 days of
`next_review_date` (as seen in the `reasons` blob for
`OVERDUE_FOR_HEARING`). Those who are in Restrictive Housing but not in
either of the other categories will be surfaced as well.

Includes a new field `eligibility_category` for analysis and validation,
with values "OVERDUE", "UPCOMING", and "OTHER", corresponding to the
categories described above.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#21089

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3c9406a7a442ba83ee1e29e3225fcc7d64876af7