Adds parole hearing count to incarceration aggregated metrics (Recidiviz/recidiviz-data#21451)
## Description of the change

This PR does the following:

- Adds `PAROLE_BOARD_HEARING_COUNT` to aggregated_metric_view_collector
for the INCARCERATION population. This is a count of the number of
parole hearings, with an attribute of `decision`-- whether the parole
hearing was approved/denied/continued.
- creates new `PersonEventType` for `PAROLE_HEARING`

I'm hoping to add 1-2 more attributes.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#21433

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Zack Goodman <zackg@recidiviz.org>
GitOrigin-RevId: b4181cc1534a27e7c510b2f9fe7f0cdc47c6958e