[US_PA] Add raw file config for another supervisor roster (Recidiviz/recidiviz-data#21663)
## Description of the change

We received a roster from PA that includes the hierarchy of leadership
in PBPP from unit supervisors to regional managers. I want to use this
information in `state_staff` views to make sure that regional managers
and district directors can be identified for outliers reporting, so this
PR adds a raw data config file to ingest the spreadsheet as a reference
file.

The spreadsheet was originally a PDF that got emailed to Lily - I
formatted it by hand in a google sheet, which I noted in this config
file. It would be ideal to receive an updated version of this roster as
a CSV regularly, but for now this is what we have to work with. I
gleaned the supervision region information from the existing
`location_metadata` table in PA; the rest of this information was
included in the original PDF.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9cc454cafca9319e96e88d2453ddcef6c3643f07