[Airflow] Update `KubernetesPodOperator` `image_pull_policy` (Recidiviz/recidiviz-data#21670)
## Description of the change

The new view update task failed during the post-deploy staging run.
There were 2 failed deploys that tagged our appengine image with
`v1.523.0`. I'm not sure what process ran in Airflow that caused
Kubernetes to pull this image , but the image from the first failed
deploy was cached in the cluster.

The failure logs show that the image used when running the view update
task was the image built from 89d49f2. The latest version includes the
new entrypoint.

```python
[2023-06-21, 02:34:25 UTC] {taskinstance.py:1770} ERROR - Task failed with exception
Traceback (most recent call last):
...
  File "/opt/python3.8/lib/python3.8/site-packages/airflow/providers/cncf/kubernetes/operators/kubernetes_pod.py", line 673, in cleanup
    raise AirflowException(
airflow.exceptions.AirflowException: Pod update-all-managed-views-3b1u24x6 returned a failure:
...
remote_pod: {'api_version': 'v1',
 'kind': 'Pod',
            'container_statuses': [{'container_id': 'containerd://9b9cfd30c084083328445e57ea9c6d126c50bfb027f696dd5d229fb0494e1ff8',
                                    'image': 'us.gcr.io/recidiviz-staging/appengine/default:v1.523.0',
                                    'image_id': 'us.gcr.io/recidiviz-staging/appengine/default@sha256:3b549863f0be5e5b5d9d6044d7cc27d4f72eaa9de9913753fc21e9f600635e75',
```

This updates the `KubernetesPodOperator` to always query the container
registry to see if the latest tagged image is cached. If it is not, it
will pull it.

https://kubernetes.io/docs/concepts/containers/images/#image-pull-policy
>The caching semantics of the underlying image provider make even
imagePullPolicy: Always efficient, as long as the registry is reliably
accessible. Your container runtime can notice that the image layers
already exist on the node so that they don't need to be downloaded
again.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6a5762ab4667ca79125adc33c5ce7e7d4f4f4375