fix Outliers merge conflicts (Recidiviz/recidiviz-data#21713)
Looks like some overlapping merges led to various Outliers report
fixtures drifting out of sync with their types.

For the change in `OfficerMetricEntity`, since the newly added attribute
is not used for the supervisor reports anyway, it seems reasonable to
just make the argument optional.

For the fixtures used in the querier tests, the event name had to be
added to the expected values now that it is required for all metrics.

GitOrigin-RevId: 9f6c4598c43df42d5d43b00591058919bbdf8bd7